[frontend] display messages and usernames together
Move ChatRoom body to app/ui/room/chat-room.tsx
