+++ frontend/app/room/[id]/page.tsx
@@ -26,121 +26,22 @@
 //  );
 //}
 
-"use client";
+import ChatRoomCard from '@/app/ui/room/chat-room';
+import { getUserId } from '@/app/lib/session';
+import { getUser } from '@/app/lib/actions';
 
-import {
-  Card,
-  CardHeader,
-  CardContent,
-  CardFooter,
-  CardDescription,
-  CardTitle,
-} from "@/components/ui/card";
-import { Input } from "@/components/ui/input";
-import { Button } from "@/components/ui/button";
-import { ScrollArea } from "@/components/ui/scroll-area";
-import ChatHome from "@/app/ui/room/chat-room";
-import { useState, useEffect } from "react";
-import { socket } from "@/socket";
-
-type Chat = {
-  text: string;
-};
-
-type MessageLog = Array<Chat>;
-
-export default function ChatRoomPage({
-  params: { id },
+export default async function Page({
+  params: { id }
 }: {
-  params: { id: number };
+  params: { id: number }
 }) {
-  const [message, setMessage] = useState("");
-  const [messageLog, setMessageLog] = useState<MessageLog>([
-    {
-      text: "example message logs",
-    },
-    {
-      text: "Hello world",
-    },
-    {
-      text: "hoge hoge",
-    },
-  ]);
-
-  useEffect(() => {
-    const newMessageReceived = (e: any) => {
-      console.log(`received message: `, e);
-      setMessageLog((oldMessageLog) => [...oldMessageLog, { text: e }]);
-      console.log(messageLog);
-    };
-    socket.on("sendToClient", newMessageReceived);
-    return () => {
-      console.log(`return from useEffect`);
-      socket.off("sendToClient", newMessageReceived);
-    };
-  }, [messageLog]);
-
-  useEffect(() => {
-    socket.connect(); // no-op if the socket is already connected
-    socket.emit("joinRoom", id);
-    console.log("emit joinRoom");
-
-    return () => {
-      socket.emit("leaveRoom", id);
-      console.log("emit leaveRoom");
-      console.log("disconnect");
-      socket.disconnect();
-    };
-  }, []);
-
-  const sendMessage = async (e: React.SyntheticEvent) => {
-    e.preventDefault();
-    const newMessage = message;
-    console.log(`sendMessage`, newMessage);
-    socket.emit("newMessage", newMessage);
-    setMessage("");
-  };
-
-  return (
-    <div className="flex items-center justify-center">
-      <Card className="w-[800px] grid grid-rows-[min-content_1fr_min-content]">
-        <CardHeader>
-          <CardTitle>Chat room</CardTitle>
-          <CardDescription>Experimental chat room</CardDescription>
-        </CardHeader>
-        <CardContent>
-          <ScrollArea className="h-[500px] w-full space-y-1 pr-3">
-            <div className="flex gap-3 text-slate-600 text-sm">
-              <ul>
-                {messageLog.map((message, i) => {
-                  return (
-                    <div key={i} className="flex gap-3 text-slate-600 text-sm">
-                      {" "}
-                      <li>{message.text}</li>
-                    </div>
-                  );
-                })}
-              </ul>
-            </div>
-          </ScrollArea>
-        </CardContent>
-        <CardFooter>
-          <form
-            className="space-x-2 w-full flex gap-2"
-            id="chat-content"
-            onSubmit={sendMessage}
-          >
-            <Input
-              placeholder="Message..."
-              name="message"
-              value={message}
-              onChange={(e) => setMessage(e.target.value)}
-              type="text"
-            />
-            <Button type="submit">Send</Button>
-          </form>
-        </CardFooter>
-      </Card>
-    </div>
-  );
+  const currentUserId = await getUserId();
+  console.log('userId: ', currentUserId);
+  if (!currentUserId) {
+    console.error("error");
+    return null;
+  }
+  const currentUser = await getUser(parseInt(currentUserId));
+  console.log('username: ', currentUser.name);
+  return <ChatRoomCard id={id} user={currentUser} />;
 }
